Rename package and module
Package: -> freebox-api
Module: -> freebox_api
