+++ src/freebox_api/.gitignore
