+++ src/freebox_api/api/.gitignore
