+++ src/freebox_api/api/connection.py
