+++ src/freebox_api/api/home.py
