+++ src/freebox_api/api/lan.py
