+++ src/freebox_api/api/lcd.py
