+++ src/freebox_api/api/netshare.py
