+++ src/freebox_api/api/notifications.py
