+++ src/freebox_api/api/storage.py
