+++ src/freebox_api/api/switch.py
