+++ src/freebox_api/api/system.py
