+++ src/freebox_api/api/wifi.py
