+++ src/freebox_api/constants.py
