+++ src/freebox_api/exceptions.py
