+++ tests/test_version.py
@@ -1,11 +1,11 @@
 """Test the single source version in pyproject.toml"""
 from packaging.version import Version
 
-import aiofreepybox
+import freebox_api
 
 
 def test_single_source_version() -> None:
     """
     It is compliant with the standard version scheme for Python packages
     """
-    Version(aiofreepybox.__version__)
+    Version(freebox_api.__version__)
